+++ tests/typing/typing_termui.py
@@ -1,0 +1,7 @@
+from pathlib import Path
+
+import click
+
+path = Path("example.txt")
+click.edit(filename=path)
+click.edit(filename=[path, "other.txt"])