+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 Unreleased
 
+- `edit()` accepts path-like objects for its `filename` parameter.
+  {issue}`2869`
 - Supported versions of Windows enable ANSI terminal styles by default.
   Colorama is no longer a dependency and is not used. {issue}`2986` {pr}`3505`
 - {class}`Argument` accepts a `help` parameter, and help output includes